fix(shared): decimal OTP durations, haversine clamp, image URL encoding docs
OTP's Leg.duration is a Float — wire durations are now Double.
Haversine clamps its intermediate to [0,1] against NaN at antipodes.
Contributing/ADR wording no longer misleads REUSE tooling or cites
internals.

+++ shared/src/commonMain/kotlin/it/iterapp/core/repo/PlanRepository.kt
@@ -43,7 +43,7 @@ class PlanRepository(
 internal fun OtpItinerary.toDomain(): Itinerary = Itinerary(
   startMs = startTime,
   endMs = endTime,
-  durationSeconds = duration,
+  durationSeconds = duration.toLong(),
   numberOfTransfers = numberOfTransfers,
   walkDistanceMeters = walkDistance,
   legs = legs.map { it.toDomain() },
@@ -79,7 +79,7 @@ internal fun OtpLeg.toDomain(): Leg {
     to = toWp,
     startMs = startTime,
     endMs = endTime,
-    durationSeconds = duration,
+    durationSeconds = duration.toLong(),
     distanceMeters = distance,
     isRealTime = realTime,
     arrivalDelaySeconds = arrivalDelay,
